requirements.txt
aiohttp==3.10.5
grpcio==1.66.1
grpcio-tools==1.66.1
openai==0.27.8
protobuf==5.28.0
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.31.0
scikit-learn==0.24.2